+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.tasks.*;
+
 import java.util.ArrayList;
 
 public class TaskList {
