+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.TaskList;
+
 public abstract class Command {
 
     public abstract void execute(TaskList taskList) throws DukeException;
