+++ src/main/java/duke/tasks/Event.java
@@ -1,5 +1,7 @@
+package duke.tasks;
+
 /**
- * A Task that start at a specific date/time and
+ * A duke.task.Task that start at a specific date/time and
  * ends at a specific date/time.
  *
  * @author Lethicia
@@ -9,7 +11,7 @@ public class Event extends Task{
     protected String at;
 
     /**
-     * Constructor for an Event task.
+     * Constructor for an duke.task.Event task.
      *
      * @param description the title or description for the task
      * @param at date/time of the task.
@@ -20,7 +22,7 @@ public class Event extends Task{
     }
 
     /**
-     * Constructor for an Event task.
+     * Constructor for an duke.task.Event task.
      *
      * @param description the title or description for the task
      * @param at date/time of the task.
