+++ src/main/java/duke/tasks/taskType.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 public enum taskType {
     TODO,
     DEADLINE,
